+++ libraries/compute-utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # compute-utils changelog
 
+## 1.43.9 (2025-06-09)
+
+- Updated UI/UX in the Sensitivity Analysis view
+- Implemented the MOEA/D method for multi-objective optimization
+- Added the optimization view
+
 ## 1.43.8 (2025-06-03)
 
 Compute2 help improvements